Support both reqwest's Request and RequestBuilder
We want to support `Request` because it can be created separately
to a specific client, but we also would like to support
`RequestBuilder` since the `send` command skips exposing the
`Request`.

Adding a separate trait for `configure` that is generic on the
"builder" type allows it to work for both types.