Fetch owned credentials for an operation
To support credentials that rotate, use a fetched object to own the
credentials for the duration of the operation.  This will usually be an
`Arc` containing a specific set of credentials, even when the
`Credential` object has since had its credentials rotated.

Also propagate `step()` into credential to allow it to pause execution
for all threads while waiting for credential to be renewed.